:handshake: Pair refacto and review
