+++ .gitignore
@@ -35,3 +35,5 @@ yarn-error.log*
 *.tsbuildinfo
 
 .slicemachine
+.now
+.vercel
