+++ tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "@tsconfig/next/tsconfig.json",
+  "compilerOptions": {
+    "strict": true,
+  },
   "include": [
     "next-env.d.ts",
     "**/*.ts",
